pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "viviian"
version = "0.1.0"
description = "VIVIIan API for hardware-agnostic data sampling"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyarrow",
    "numpy",
    "toml",
    "scipy",
    "networkx",
    "pythusa",
]

[project.optional-dependencies]
ni  = ["nidaqmx"]
gui = ["imgui", "glfw", "PyOpenGL"]

[tool.setuptools.packages.find]
where = ["src"]
